requirements.txt
tensorflow-cpu
requests
numpy
pillow